Don't let the album applet freeze Amarok for ages
We tried to show ALL tracks from ALL artists, that do not belong to an
album, if some track from the currently playing artist did not have an
album set. Took about 50 seconds for me on track change...

TODO: show all tracks from this artist that are not part of an album.

BUG: 277021
BUG: 260810
FIXED-IN: 2.4.2