Fix crash in WikipediaApplet.
The crash happened because of the loadProgress() signal triggering the
_pageLoadProgress() slot in the Wikipedia applet *after* the
_pageLoadFinished() has been triggered already.
That caused a null-pointer dereference.

The actual problem is that we are not disconnecting from the webView
loadProgress() signal after some page loading has been finished.

Also remove some not-so-pretty static_assert statements and use a more
type-safe approach to keep track of the progress bar proxy widget.
Use some soft asserts to avoid the worst-case when the proxy widget may
have been deleted somehow.

BUG: 279813
FIXED-IN: 2.5