Rework MediaDeviceAlbum cover-related functions, remove "Unset Cover"
There was a logic error in MediaDeviceAlbum::image() method which was
the cause of bug 263268, this is fixed and cover setting functions are
cleaned-up.

The "Unset cover" MediaDeviceAlbum action never worked, so remove it
from GUI and remove related stub methods. I believe the functionality
is so little used on media devices so it is not worth implementing it
(yet).

BUG: 263268
FIXED-IN: 2.5