MediaDeviceTrack: prevent possible crashes, Handler::Capability fixes
When a media device is removed, its collection and handler are
destroyed, but its tracks, albums are kept as long as they are
referenced. Tracks, albums use QWeakPointer to track deletion of
their handler. Ensure that this pointer is always checked before
use in MediaDeviceTrack.

Also ensure that related Capabilities get destroyed (they should
be alredy tracked using QWeakPointer) when handler gets deleted.
It did not work previously in cases when Capabilities were created
in a thread from handler's. Move them to handler's thread so that
parentship can be established.

BUG: 284838
FIXED-IN: 2.5