SqlQueryMaker: likeCondition should be case-insensitive in all cases
When creating a SQL "LIKE" condition and both the beginning and the
ending of the string have to match, the comparison was case sensitive.
In all other cases (only one or no end had to match) the comparison was
case sensitive.

This was too convoluted so this commit changes it to be always
case-insensitive. This case was probably never hit until
https://git.reviewboard.kde.org/r/102252/ was merged.

REVIEW: 103015