Do not crash in Hander::Capability constructor in Qt debug builds
Hander::Capability::Capability() used to have following code:
moveToThread( hander->thread() );
setParent( handler );

This unfortunately asserts with: ASSERT failure in QCoreApplication
::sendEvent: Cannot send events to objects owned by a different thread
in debugging Qt builds.

Put setParent() call into a slot, and emit a signal to trigger it so
that it is called in object's new thread.

This should fix bug 279798, but it cannot be tested in current master
until bug 285720 is resolved.

BUG: 279798
FIXED-IN: 2.5