Maintain authors list.
Add Myriam, Sven and Ralf and remove people who've been MIA since > 1.5 years.

Authors status is granted to long term contributors and comes with a
the assumption that they will stay active for a long while and stay
invested in the maintenence and/or expansion of the project.

CCMAIL:amarok-devel@kde.org