Fix infinite loop in context view eating CPU until main window is shown
We used to use _scene_ sceneRect in
VerticalToolbarContainment::updateGeometry() to update applets and
geometry, but that leaded to infinite loop (across mainloop) -
m_applets->setGeometry(), refresh() would enlarge _scene_ sceneRect by
a few pixels which would trigger updateGeometry() and so on...

We now use _view_ sceneRect to update geometry. Rick W. Chen discovered
some glitches in applet collapsing when this patch is applied, but he
has a patch for that in the works and this is more serious problem, so
pushing this.

REVIEW: 103271
BUG: 278897
FIXED-IN: 2.5