media device collection: more consistent handling of compilations
This makes handling of compilation albums consistent across:
 * Amarok collection
 * iPod viewed in Amarok
 * iPod viewed on itself

The code used to determine best guess album artist in
ScanResultProcessor was factored into ArtistHelper for reusability,
ArtistHelper is now exported in amaroklib.

This should definitely fix bug 232072. (or better - I'm not able to
reproduce it with this patch applied)

BUG: 232072
FIXED-IN: 2.6
DIGEST: compilation handling on iPod improved