Collection browser: show absolute free space on hover, thicker bar
For collections that support capacity (iPod, UMS, MTP), change text on
capacity bar from "80% used" to "1,2 GiB free". Rationale:
 * relative emptiness is already indicated using bar
 * absolute free space is the most relevant information when
   transferring tracks
 * this is consistent with Dolphin

Additionally, the capacity bar is made thicker and the font is non-bold
so that the font is readable when rendered on it. (this is done by
using KCapacityBar's minimumSizeHint() - however it is bounded between
MIN and MAX values)

This may be a little controversial change of the look, here are some
screenshots, speak up if you don't like it:
 * before: http://troll.ws/i/VanuSm.png
 * after: http://troll.ws/i/RTSrAq.png

DIGEST: feature: more useful capacity bar when hovering collection