MemoryMeta::MapChanger: implement trackChanged()
We use a little trick to remove the track from MemoryCollection and
then add it back, but only when the change was big enough that this
is needed. (a change to track comment really don't need to be reflected
to MemoryCollection maps)

Changing album artist is currently a bit funny, because it changes
album artist for all tracks in album. This is how MemoryMeta currently
works, albums are identified only using their name.