MetaFile::Track: simplify isPlayable(), isEditable(), fix tests
QFileInfo is a nice class, use it. The testmetafiletrack failed when
run as root, because it sets permissions of a test file to 0x000 and
then asserts on !track.isEditable(), but for root every file is
editable regardless of its permissions. Skip this one assert is such
case. (when the file is writeable even with no permission bits set)