Update playlist selection at appropriate time when text changed in search
When "show only matches" is enabled the listview tries to select the track
and emit a signal about whether or not it matched before the proxy timeout
has triggered the update. This happens to work anyway when additional
keystrokes are narrowing the results but causes problems when I typo. For
example, in my play list, "laz" returns 15 results and "lazz" returns 0.
If I then backspace the second "z", it doesn't select the top track and
the lineedit widget still appears in red.

This is a fairly direct solution: moving the relevant selection code to
the timeout handler with some member vars to store the required arguments.
If there's a better way to go about it please let me know and I'll try to
bring it up to scratch.

Author: Thomas Karpiniec

REVIEW: 101974