Fix crash when remounting UMS.
Fix crash when a device holding a UmsCollection is mounted again after an unmount without removing the device in between.
The crash occours because the QSharedPointer m_mc is cleared during an unmount but not initialized when mounted again. I fixed this issue by initializing m_mc in the init-method and not in the constructor.

Patch by Volker Christian

BUG:289751
FIXED-IN:2.6

REVIEW:103774