Change GUI strings in Configuration dialog to reduce user confusion
This is a result of a dialogue with an Amarok user (Leonidas Tsampros)
on #amarok who (logically) thought that if he enables "NFS Device" in
"Devices" section of Amarok plugins, NFS will pop-up as another
collection.

This is not how SMB, NFS, MassStorage "Devices" work currently, they
are meant to enable users to have a part of their collection on
respective devices.

I believe that the old plugin names and descriptions are confusing to
the extent that their translations should be dropped; it should be a
plenty of time till Amarok 2.6 so translators should have opportunity
to pickup these. (or, do .desktop translations get erased automatically
when original string changes?)

CHANGES:
 * "Devices" in Amarok configuration -> Plugins is with other related strings
   renamed to "Local Collection Backends" to reduce user confusion.

BUG: 293277
FIXED-IN: 2.6
REVIEW: 103871