UmsCollection: rework factory <-> collection interaction
The main change is that the UmsCollection is created only for
_accessible_ (mounted) Solid devices so no phantom (allowing no
interaction) collections are listed in collection browser.

Another change is that not-activated UMS collections get
"(not activated)" appended to their name so that it is more clear that
users need to click "Use As Collection" button, which is by the way
renamed to "Activate This Collection" by this commit.

The changes in UmsColllectionFactory are done with future merge with
other "new" pluggable collections factories in mind -- 80% of the code
in UmsColllecionFactory is the same as in out-of-tree rewrite of the
iPod collection.

CHANGES:
 * Don't show unmounted USB Mass Storage devices and make it clear when
   the device is not "activated" yet.

CCMAIL: Bart Cerneels <bart.cerneels@kde.org>
DIGEST: Better behaviour of USB Mass Storage collection [see CHANGES]