avoid code duplication in user playlists; fixes to playlist browser
UserPlaylistProvider subclasses now need not to provide their own
delete playlist, rename playlist and remove track from playlist
actions, these are provided by base class now.

Rename playlist action is changed to be an in-line rename, plus
PrettyTreeView is modified to assure that an item that is going to be
edited has its parents expanded.

Inline renaming is additionally made more robust in order not to rename
a different playlist with the same name.

Renaming of playlists from is model is changed to go though PlaylistManager
and respective Provider - providers may want to do more that just call
playlist->setName().