Fix invalid assertion in PlaylistBrowserView::slotActivated()
The code assumed that if the item action list is non-empty, its first
action must be "appendAction". This is not true, for example for root
items (PlaylistProviders) that can have custom actions.

We now don't activate any action in slotActivated() if the first action
is not named "appendAction".

CCMAIL: Bart Cerneels <bart.cerneels@kde.org>