move the Free Music Charts script from playground/ to src/
That feature is part of VLC by default and is going to be included in
Tomahawk soonish, too.

The script is very well tested, very popular and received a lot of
positive feedback during the last few years. So I don't see any reasons
not to include it. Don't prove me wrong. ;-)

Also there is a massive copyright debate going on in Germany and some
free music artists complained to me, that it is nearly impossible to get
attention for their music, as traditionial media are more or less dominated
by the large major labels. So let's help those artists. :)

DIGEST: Ship the Free Music Charts script with Amarok by default