QtGroupingProxy: don't re-enter beginInsertRows() (crashfix)
This should fix a nasty crash that was only reproducible when Qt was
build with debugging assertions enabled. (and perhaps only if you have
empty playlist providers) The fix is based on an assumption that
addSourceRow() in QtGroupingProxy is only called from:

a) buildTree(), which encompasses it in {begin/end}ResetModel()
b) modelRowsInserted(), but that should be always preceeded by
   modelRowsAboutToBeInserted() that calls beginInsertRows()

Bart, this is largely your code, is this correct?

BUGFIXES:
 * Fix crash on start when Qt is build with debugging assertions

BUG: 285720
FIXED-IN: 2.6
DIGEST: fix rather nasty crash