Amazon store: ability to check single items out directly
So the user does not have to use the integrated shopping cart (which
does not work for Amazon.com at all, due to their API limitations).
This currenly only works correctly for Amazon.com, some server side
changes are required for the other locales. Those will show up shortly.

BUG:288808
FIXED-IN:2.6