Collection browser: don't add duplicate tracks in mimeData
That could have happened when you selected e.g. both album and a track
from that album -> the track would be added twice.

The implemented algorithm is done at QModelIntex level and is O(n)
where n is the number of selected tracks so this should be okay.

There is an unrelated bug with the Various Artists item that behaves
as entire collection when dragged. This will be fixed soon.

BUGFIXES:
 * Don't add duplicates to playlist when dragging from collection

BUG: 254411
FIXED-IN: 2.6
DIGEST: Bugfix: Amarok now doesn't add unwanted duplicates to playlist