StatSyncing: move to amaroklib, back-end for collections, track matching
This commit implements following changes/features:
 * bury amarok-statsyncing.so: because CollectionManager in in amaroklib,
   there would be circular dependency which is not possible. StatSyncing
   is now part of amaroklib and built from src/CMakeLists.txt
 * a bridge between Amarok Collections, Meta and StatSyncing is
   implemented in statsyncing/collection. This means that StatSyncing can
   "see" collections and collection tracks, yay!
 * TrackDelegate comparisons are overhauled to be more intuitive to use
   and flexible because they accept comparison mask now
 * MatchTracksJob is implemented and called by Synchronize Statistics...
   action, it will match tracks from all visible collections

How to test: plug in all devices that show up as collection in Amarok,
start Amarok in console as amarok --debug. Hit "Synchronize Statistics...
from the Tools menu and watch all the beauty that appears in debugging
log. Notice the UI is still responsive while matching because it happens
in a background thread. I'm interested in first and last screen of that
debugging output, you may pastebin it and ping strohel at #amarok (IRC)
about it. Thanks. :-)