Amazon store: try to show a sensible default in the country selection
If the user has not set a location we now try to guess one by asking
KLocale for the country setting. This is then being displayed as default
choice in the country selection dialog.

This does NOT yet save this determination or even use it to query the
service, as sending data to an internet service without explicit
consent from the user is a bad, bad thing. It just sets the default
value in the dialog the user still has to confirm.