IpodCollection: prevent crash on early disconnect and double-freeing
We were deleting Itdb_Playlist's twice, once in itdb_free and for the
second time in IpodPlaylist::~IpodPlaylist. Ensure we delete it only in
the second call by using trick we already do with tracks.

Also, try to prevent crashes when IpodCollection is destoryed while
IpodPlaylistProvider is still in parseItdbPlaylists() (needs testing)

CCBUG: 300958
FIXED-IN: 2.6