Meta::MetaCapability: sanitize and document interface
MetaCapability interface is divided by this commit into:
 * public: has<CapIface>() and create<CapIface>()
 * "protected": hasCapabilityInterface() and createCapabilityInterface()

The protected methods are in fact public, because various proxy classes
need to call them on other instances, protected inheritance doesn't
allow this.

Also document that the caller of create<CapIface>() needs to delete it
afterwards.

Convert existing (non-proxy) callers of hasCapabilityInterface() to
has<CapIface>() to promote usage of the "public" API.