Fix inconsistent scrobbling with GStreamer backend
With the GStreamer backend, the trackChanged signal seems to be emitted
twice on track change, which in turn sometimes triggered a consistency
check in the scrobbling code and sometimes causes tracks not to be
scrobbled. About 1/5 (variable) tracks would be successfully scrobbled.

This should not affect scrobbling with the VLC backend, which emits
correct signals.

This is a temporary hack until Matěj's statistics rewrite is complete.