TestPlaylistFormat: leverage data-driven testing, formatting fixes
Data-driven testing is used in testGetFormat() to make the test
shorter, more elegant and less repetitive. Jasneet, you may want to do
the same for testIsPlaylist(), I'm too lazy. :-)

Additionally, tabs are converted to spaces. Including .moc file is not
needed unless Q_OBJECT is in a .cpp file, so it is removed.

CCMAIL: Jasneet Bhatti <jazneetbhatti@gmail.com>