AmarokPlaylistScript: play all media in {add,play}Media(), add playMediaList()
No need to convert to tracks (this way it excluded playlists),
PlaylistController is good enough to read all kinds of urls including
playlists.

Add playMediaList() method for symmetry with addMediaList() when at it.

We thought it worked for other than .m3u playlists, but it was by
chance becasce CollectionManager::trackForUrl() returned something
(probably the first track) for other playlists.

BUG: 232697
FIXED-IN: 2.6