Background job of NepomukCollection added
The query and enumeration runs as a background job in
NepomukCollection. But the UI doesn't show the tracks
properly. I suspect this is because the UI gets updated
even before the tracks are queried and enumerated. There
should be a mechanism to push the queried results in batches.
This should solve this issue