Resolved graying out of tracks
Still empty NepomukAlbum are constructed when a track with no album is
encountered. This might be the reason for the numerous 'Unknown Album'.
The reason for not fixing this is, there might be a bug in
MemoryMeta::MemoryChanger::addExistingTrack() which doesnt handle
empty albums right. It is breaking the code.