SPARQL query implemented instead of Nepomuk::Resource API
Hefty changes in the core of the background job of the query of tracks
and subsequent enumeration.

Initially a manual sparql query which queries for all properties of the
track and other meta data like artist, album etc.

The meta HashMaps are changed to incorporate Resource URIs and Labels
instead of resources themselves, as keys.

Year and labels ( tags in nepomuk ) are yet to be implemented