Add configuration as capability action
Because collections don't have configuration dialogs like services, so
it's now added as a capability action, the collection will be
registered
immediately after it's initialized, so Spotify collection is always
shown in collection manager as long as the plugin is activated.

Automatically download Spotifu resolver from specified server when
it's
not found in local system.

Remove legacy code.