FileType: don't translate all file types, just the "Other"
This fixes a bug where tracks cannot be copied to iPods in de locale
because mp3 is apparently translated to MP3 there and IpodCollection
compares it case-sensitively against a list of supported file types,
which fails.

For easier back-portability, this patch includes only the actual fix,
second part with misc changes (like ChangeLog and docstrings) will
follow.

BUG: 305134
FIXED-IN: 2.7