Make Amarok compile and link with TagLib installed in $HOME
I installed TagLib 1.8-GIT in $HOME so I can try the new features (see also my other review request #101598). $HOME/bin is in $PATH and thus "taglib-config" returns all the right settings. Still Amarok did not compile. The reason is that it adds the taglib include directory to the list of include directories, but then still accesses the taglib headers with "#include <taglib/tstring.h>" in some places (instead of using "#include <tstring.h>"). This is clearly wrong, but coincidentally
works if taglib is installed in /usr. I fixed that.

Note: the taglib linking change was not included as discussed on reviewboard.

Courtesy: Mathias Stephan Panzenböck

FEATURE: 104693
FIXED-IN: 2.7