{Smb,Nfs}DeviceHandler: use Solid instead of KMountPoint
KMountPoint seems to be unreliable and not working at least for smb
shares. Use Solid's NetworkShare for the same task, which is much more
*cough* ... solid. :-)

Server name & share name extraction changed a bit, but I've tested it
to give the same names as the old implementation. (not for some corner
cases, but old behaviour could be considered bogus there)

v2: use QUrl::host(), ::path() instead of uglier QString::section(),
reorder some calls for more meaningful debug messages.

This does not yet solve bug 302837 fully, just the first part of it.

CCBUG: 302837
FIXED-IN: 2.7
REVIEW: 106094