Amazon store: give the user hints to a simpler checkout
Amazon asks the user to download their downloader as long as the
"download installed" cookie is not set in the browser. The checkout
process is hard to compelte without that cookie (you have to find quite
well hidden links to continue).
Therefore I'm putting links to set that cookie at several places in
the gui.