Make Meta::MetaBase::(un)subscribe() non-virtual, clean subclasses
It is extremely unsafe not to use MetaBase::m_observers, to prevent
everyone from doing so. MetaProxy needlessly implemented its own
mechanism, ditch it.

This avoids many potential problems and one actual: testmetamultitrack
was failing for me because of custom MetaProxy observer implementation.

This also removes ~70 lines of code, nice.