On Linux, try to set the scanner's IO priority instead of CPU priority.
If the CPU was always busy the scanner would never run. This isn't a
theoretical scenario - distributed computing clients like BOINC will
keep the CPU busy with minimal impact to the user.
Always keeping the disk busy is something no program should do on a
machine used interactively, so that the collection scanner should
usually get a chance to run. Not stalling IO of more important things
is also the actual goal of setting the scanner's priority AFAICS, so
setting IO priority is a good thing all around.