Amazon store: Support for Amazon MP3 in Italy and Spain
Checkout does not work yet, we need to be verified by Amazon first. That
should be a matter of a few days and does not require any changes in
Amarok.

So now we support: USA, Great Britain, France, Spain, Italy, Japan,
Germany, Austria and Switzerland

We might support Andorra, Vatican and Portugal too, with the new
locales. So if anyone living in those countries could test the store
(once the new ones have been verified by Amazon) it would be great.

I always wanted to hear some words from the pope about free software. ;-)

BUG:307981
FIXED-IN:2.7
DIGEST: Support for Amazon MP3 in Spain and Italy