QueryMaker: fix flawed API wrt ArtistQueryMode
ArtistQueryMode was not really a "mode", just a switch for
addMatch( Meta::ArtistPtr ) behaviour.

Rename it to ArtistMatchBehaviour, remove setArtistQueryMode() method
and instead add a parameter to addMatch( Meta::ArtistPtr, .. ) method
defaulting to TrackArtist as in past.

CCing amarok-devel because this is a rather intrusive API change.

Phalgun, this should help you with NepomukQueryMaker as the interface
makes more sense now.

CCMAIL: amarok-devel@kde.org