Document that strongly recommended liblastfm version in 1.0.3
We tell users that we need 1.0.3, but we really check just >= 1.0.0.
This is because upstream forgot to update version in lastfm/global.h,
so it looks like 1.0.2. :-( We happily build with 1.0.0 - the bugfixes
are runtime and we get them even if built agaist older liblastfm.

We need (recommend) 1.0.3 for 2 bugfixes:
 * playcount support in scrobbling (from iPod)
 * bugfix in tag-removing webservice procedure making it work