QtGroupingProxy: implement buddy() to work-around design issues
This is effectively a work-around for a big design flaw in
QtGroupingProxy (which is that it "invents" its own items not present
in the original model). Technical description in the code comments.

^^^^ Bart, because of the above reason I'd prefer not to use
QtGroupingProxy in Amarok at all in long-term. Instead, we should do it
the other way around: original model would be hierarchical and the
proxy would flatten it into a table.

This fixes some bugs in the Saved Playlists and perhaps more:

BUGFIXES:
 * Fix editability and drop-ability of playlist folders.

CCBUG: 308703
CCMAIL: Bart Cerneels <bart.cerneels@kde.org>