TagHelper: provide meaningful default implementation of render()
render() is used to calculate unique id of a track that doesn't have it
embedded. Providing default implementation means that track metadata of
more obscure formats (MP4, ASF) will have the metadata contained in
the unique id hash in addition to file size and first 16 KB of a file.

This can change hashes of existing files, but thanks to recent fixes,
Amarok can cope with it and update them as needed. (even in playlists)

CCBUG: 258555