ActionClasses, KNotificationBackend: announce stop after track on/off
This makes KNotificationBackend more usable so that it can replace OSD.
This change unfortunately introduces new translatable string. I haven't
found a way to do without, sorry about this, but I consider this an
important bugfix and the new string is in fact not really accessible.
(only in Configure Notifications, however Amarok doesn't offer this
button - it is only available through System Settings)