Fix The size of rating stars should not change in inline editing
Improved a lot of size issues. You need to pull a lot of tricks out
of your bag to get the widgets to have exactly the same layout.
Still it seems that I am missing a pixel or so somewhere.

Changing the ratings directly with one click is another issue that
I don't want to tackle right now.

BUG: 209684
FIXED-IN: 2.7