Update Dynamic playlist. Get rid of energy optimizing code
The dynamic playlist always used to optimize the generated tracks
by different ways.
However the new biases all have hard conditions. The user also
used to see it like this and was confused if a "pretty good" result
was generated.

The new algorithm will generate an empty list if the conditions
cannot be fullfilled

BUG: 286642
BUG: 292648
BUG: 304374
BUG: 301716
FIXED-IN: 2.8