Revert special "hold Shift" context menu behaviour
CHANGES:
 * Fancy behavior of some context menus showing different actions when Shift
   key is held has been reverted. All entries are now shown all the time.

Per usability review request, this is really a bad user interaction.
This commit restores a behaviour where we show all the options:
 * Copy to Collection
 * Move to Collection
 * Move Tracks to Trash
 * Delete Tracks

Sorry for the wasted efforts, Bart & Edward.

In my opinion, we should remove the "Delete Tracks" action from the
context menu entirely, as it is really an "advanced user" action and it
will stay available as a keyboard shortcut (and, for historic reasons,
trigger-able also by Shift-clicking the "Move Tracks to Trash").

REVIEW: 108686