Refactoring: Using the same delegate almost everywhere in the browser
Fix some small issues in the CollectionTreeDelegate and moved it to
widgets as a PrettyTreeDelegate.
Also fix the strange actions rect mechanism. Now we don't have a
global static map for actions.

Create include file for PrettyTreeRoles so that all the different models
don't need to include the complete delegate or redefine the roles.
Remove the BrowserCategoryListDelegate and use the PrettyTreeDelegate