Formatting fixes to ChangeLog (only 2.8-Beta entry)
Not a big deal of course, bug 70-char long lines felt bad to my eyes. :-)

I propose we should follow following guidelines
 * Hard 90-char line length limit
 * Also don't break the early unless:
    * there would be lone "a" at the end of line
    * "(BR 123456)" would be broken to 2 lines - that would break our
       ChangeLog.html-generating script
 * Period between message and (BR 123456)
 * Patches from non-core developers should have mention the author

Alternatively to the last point, we could change it so that we always
mention the author for a bit of more credit - what do you think?

CCMAIL: amarok-devel@kde.org