Fix crashes on track change (both with Phonon-GStreamer and Phonon-VLC).
What happened:

The Wikipedia applet uses QtWebKit for HTML rendering. On
Linux, QtWebKit uses GStreamer for media playback. Turns out that this crashes
with a high probability whenever an <audio> tag is present on a web page. As
music artists' pages often contain audio snippets in Vorbis format, it crashed
a lot when Amarok's applet rendered a page.

The fix:

Removing all audio tags from Wikipedia pages before rendering them in our applet.
Of course, the real fix would have to be done upstream. But until that happens, our
workaround here fixes a shitload of crashes.

We might want to do a 2.7.1 release for Linux, with this patch.

BUG: 319371