Refactor TrackLoader
The changes in code are big, but changes in functionality are little.
New code is more careful when sorting, ensures that tracks are fully
loaded when the users needs it and replaces any blocking behaviour
with asinc code (still with strong guarantees).

Also the test is greatly improved to test nearly all the functionality.