Add analyzer applet to the user's existing config.
This trickery is needed to give the new analyzer applet some
visibility when upgrading to Amarok 2.8. We rewrite the
config carefully so that the applet gets inserted in a sane position.
We also take care not to re-add this applet, in case the user
has explicitly removed it earlier.

Testing welcome to ensure that no special cases are ignored that
could screw up the config somehow.

CCMAIL: amarok-devel@kde.org