Add tooltip to analyzer applet explaining how to configure it.
As the analyzer applet does not have a header with a configure button
like the other applets do (it would take up too much space I think),
it's not obvious how it can be configured. So I've added this tooltip
saying that you can right-click it.

Is this a good compromise?