Remove "Configure Amarok" item from the systray menu.
In a recent discussion on IRC, we agreed that having the configure
item in the tray menu does not really make sense (at least not for
Amarok). It was unclear though if it could be removed at all.

So now I found out how to remove it.

CCMAIL: amarok-devel@kde.org