[Single]CollectionTreeItemModel[Base]: first look at cleanups/corrections
This is the first part of much-needed revisiting of the Collection Browser
and associated models. The final goal is to get rid of the unfortunate
expanding/scrolling behaviour and of the caching bug. That will involve
more code floating around this commit just brings some fixes,
deduplications and corrections.

This should fix bug 262504, but needs confirmation from folks that were
able to reproduce this bug. Reporters, please reopen if you can reproduce
this with recent git or 2.8 Beta (to be released really soon)

BUG: 262504
FIXED-IN: 2.8